Fix: Enable 2 different pkgs from same repo to be inspected
Previously, if some package relied on 2 different packages in a single remote repo, go-getx would would do a clone and inspect on the first package but would take the mere existence of the second package on disk as evidence that we did not need to do an inspection. As a result, dependencies of the second package would not be retrieved and the final go install of the first package would fail.

The primary fix here is treat the fact that a package is newly cloned as a reason (despite other flags) for itself to be inspected. To achieve this I reworked some parts to split the current 'done' tracking to differentiate between 'cloned', '[go-]listed' and 'inspected'.